Add literal interpretation escape
